+++ src/tui/tui.go
@@ -91,6 +91,7 @@ const (
 	Change
 	BackwardEOF
 	Start
+	Load
 
 	AltBS
 
@@ -268,6 +269,7 @@ type ColorTheme struct {
 	Selected     ColorAttr
 	Header       ColorAttr
 	Separator    ColorAttr
+	Scrollbar    ColorAttr
 	Border       ColorAttr
 	BorderLabel  ColorAttr
 	PreviewLabel ColorAttr
@@ -304,6 +306,22 @@ const (
 	BorderLeft
 	BorderRight
 )
+
+func (s BorderShape) HasRight() bool {
+	switch s {
+	case BorderNone, BorderLeft, BorderTop, BorderBottom, BorderHorizontal: // No right
+		return false
+	}
+	return true
+}
+
+func (s BorderShape) HasTop() bool {
+	switch s {
+	case BorderNone, BorderLeft, BorderRight, BorderBottom, BorderVertical: // No top
+		return false
+	}
+	return true
+}
 
 type BorderStyle struct {
 	shape       BorderShape
@@ -429,7 +447,7 @@ type FullscreenRenderer struct {
 	mouse        bool
 	forceBlack   bool
 	prevDownTime time.Time
-	clickY       []int
+	clicks       [][2]int
 }
 
 func NewFullscreenRenderer(theme *ColorTheme, forceBlack bool, mouse bool) Renderer {
@@ -438,7 +456,7 @@ func NewFullscreenRenderer(theme *ColorTheme, forceBlack bool, mouse bool) Rende
 		mouse:        mouse,
 		forceBlack:   forceBlack,
 		prevDownTime: time.Unix(0, 0),
-		clickY:       []int{}}
+		clicks:       [][2]int{}}
 	return r
 }
 
@@ -465,6 +483,7 @@ var (
 	ColInfo                 ColorPair
 	ColHeader               ColorPair
 	ColSeparator            ColorPair
+	ColScrollbar            ColorPair
 	ColBorder               ColorPair
 	ColPreview              ColorPair
 	ColPreviewBorder        ColorPair
@@ -488,7 +507,6 @@ func EmptyTheme() *ColorTheme {
 		Cursor:       ColorAttr{colUndefined, AttrUndefined},
 		Selected:     ColorAttr{colUndefined, AttrUndefined},
 		Header:       ColorAttr{colUndefined, AttrUndefined},
-		Separator:    ColorAttr{colUndefined, AttrUndefined},
 		Border:       ColorAttr{colUndefined, AttrUndefined},
 		BorderLabel:  ColorAttr{colUndefined, AttrUndefined},
 		Disabled:     ColorAttr{colUndefined, AttrUndefined},
@@ -496,6 +514,8 @@ func EmptyTheme() *ColorTheme {
 		PreviewBg:    ColorAttr{colUndefined, AttrUndefined},
 		Gutter:       ColorAttr{colUndefined, AttrUndefined},
 		PreviewLabel: ColorAttr{colUndefined, AttrUndefined},
+		Separator:    ColorAttr{colUndefined, AttrUndefined},
+		Scrollbar:    ColorAttr{colUndefined, AttrUndefined},
 	}
 }
 
@@ -515,7 +535,6 @@ func NoColorTheme() *ColorTheme {
 		Cursor:       ColorAttr{colDefault, AttrRegular},
 		Selected:     ColorAttr{colDefault, AttrRegular},
 		Header:       ColorAttr{colDefault, AttrRegular},
-		Separator:    ColorAttr{colDefault, AttrRegular},
 		Border:       ColorAttr{colDefault, AttrRegular},
 		BorderLabel:  ColorAttr{colDefault, AttrRegular},
 		Disabled:     ColorAttr{colDefault, AttrRegular},
@@ -523,6 +542,8 @@ func NoColorTheme() *ColorTheme {
 		PreviewBg:    ColorAttr{colDefault, AttrRegular},
 		Gutter:       ColorAttr{colDefault, AttrRegular},
 		PreviewLabel: ColorAttr{colDefault, AttrRegular},
+		Separator:    ColorAttr{colDefault, AttrRegular},
+		Scrollbar:    ColorAttr{colDefault, AttrRegular},
 	}
 }
 
@@ -547,7 +568,6 @@ func init() {
 		Cursor:       ColorAttr{colRed, AttrUndefined},
 		Selected:     ColorAttr{colMagenta, AttrUndefined},
 		Header:       ColorAttr{colCyan, AttrUndefined},
-		Separator:    ColorAttr{colBlack, AttrUndefined},
 		Border:       ColorAttr{colBlack, AttrUndefined},
 		BorderLabel:  ColorAttr{colWhite, AttrUndefined},
 		Disabled:     ColorAttr{colUndefined, AttrUndefined},
@@ -555,6 +575,8 @@ func init() {
 		PreviewBg:    ColorAttr{colUndefined, AttrUndefined},
 		Gutter:       ColorAttr{colUndefined, AttrUndefined},
 		PreviewLabel: ColorAttr{colUndefined, AttrUndefined},
+		Separator:    ColorAttr{colUndefined, AttrUndefined},
+		Scrollbar:    ColorAttr{colUndefined, AttrUndefined},
 	}
 	Dark256 = &ColorTheme{
 		Colored:      true,
@@ -571,7 +593,6 @@ func init() {
 		Cursor:       ColorAttr{161, AttrUndefined},
 		Selected:     ColorAttr{168, AttrUndefined},
 		Header:       ColorAttr{109, AttrUndefined},
-		Separator:    ColorAttr{59, AttrUndefined},
 		Border:       ColorAttr{59, AttrUndefined},
 		BorderLabel:  ColorAttr{145, AttrUndefined},
 		Disabled:     ColorAttr{colUndefined, AttrUndefined},
@@ -579,6 +600,8 @@ func init() {
 		PreviewBg:    ColorAttr{colUndefined, AttrUndefined},
 		Gutter:       ColorAttr{colUndefined, AttrUndefined},
 		PreviewLabel: ColorAttr{colUndefined, AttrUndefined},
+		Separator:    ColorAttr{colUndefined, AttrUndefined},
+		Scrollbar:    ColorAttr{colUndefined, AttrUndefined},
 	}
 	Light256 = &ColorTheme{
 		Colored:      true,
@@ -595,7 +618,6 @@ func init() {
 		Cursor:       ColorAttr{161, AttrUndefined},
 		Selected:     ColorAttr{168, AttrUndefined},
 		Header:       ColorAttr{31, AttrUndefined},
-		Separator:    ColorAttr{145, AttrUndefined},
 		Border:       ColorAttr{145, AttrUndefined},
 		BorderLabel:  ColorAttr{59, AttrUndefined},
 		Disabled:     ColorAttr{colUndefined, AttrUndefined},
@@ -603,6 +625,8 @@ func init() {
 		PreviewBg:    ColorAttr{colUndefined, AttrUndefined},
 		Gutter:       ColorAttr{colUndefined, AttrUndefined},
 		PreviewLabel: ColorAttr{colUndefined, AttrUndefined},
+		Separator:    ColorAttr{colUndefined, AttrUndefined},
+		Scrollbar:    ColorAttr{colUndefined, AttrUndefined},
 	}
 }
 
@@ -634,7 +658,6 @@ func initTheme(theme *ColorTheme, baseTheme *ColorTheme, forceBlack bool) {
 	theme.Cursor = o(baseTheme.Cursor, theme.Cursor)
 	theme.Selected = o(baseTheme.Selected, theme.Selected)
 	theme.Header = o(baseTheme.Header, theme.Header)
-	theme.Separator = o(baseTheme.Separator, theme.Separator)
 	theme.Border = o(baseTheme.Border, theme.Border)
 	theme.BorderLabel = o(baseTheme.BorderLabel, theme.BorderLabel)
 
@@ -644,6 +667,8 @@ func initTheme(theme *ColorTheme, baseTheme *ColorTheme, forceBlack bool) {
 	theme.PreviewFg = o(theme.Fg, theme.PreviewFg)
 	theme.PreviewBg = o(theme.Bg, theme.PreviewBg)
 	theme.PreviewLabel = o(theme.BorderLabel, theme.PreviewLabel)
+	theme.Separator = o(theme.Border, theme.Separator)
+	theme.Scrollbar = o(theme.Border, theme.Scrollbar)
 
 	initPalette(theme)
 }
@@ -676,6 +701,7 @@ func initPalette(theme *ColorTheme) {
 	ColInfo = pair(theme.Info, theme.Bg)
 	ColHeader = pair(theme.Header, theme.Bg)
 	ColSeparator = pair(theme.Separator, theme.Bg)
+	ColScrollbar = pair(theme.Scrollbar, theme.Bg)
 	ColBorder = pair(theme.Border, theme.Bg)
 	ColBorderLabel = pair(theme.BorderLabel, theme.Bg)
 	ColPreviewLabel = pair(theme.PreviewLabel, theme.PreviewBg)